Separate deployment of token bridge creator from chain specific setup
